requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
black==23.12.0
blinker==1.9.0
click==8.2.1
factory_boy==3.3.3
Faker==37.5.3
flake8==7.3.0
flake8-bugbear==22.1.11
flake8-pie==0.15.0
Flask==3.1.1
flask-openapi3==4.2.1
flask-openapi3-elements==9.0.6
flask-openapi3-rapidoc==9.3.8
flask-openapi3-rapipdf==2.2.1
flask-openapi3-redoc==2.5.0
flask-openapi3-scalar==1.34.2
flask-openapi3-swagger==5.27.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==5.9.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.6.3
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
py==1.11.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
SQLAlchemy==2.0.42
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
